feat: tests
add a test framework

+++ obsidian_parser/multiplication.py
@@ -1,10 +1,10 @@
 class Multiplication:
     """Multiply number by multiplier."""
-    def __init__(self, multiplier: int):
+    def __init__(self, multiplier: float):
         """Initialize multiplier."""
         self.multiplier = multiplier
 
-    def multiply(self, number: int) -> int:
+    def multiply(self, number: float) -> float:
         """Multiply number by multiplier."""
         return number * self.multiplier
